chore: components restructuring and minor UI improvements.
