+++ packages/types/src/issues/base.d.ts
@@ -10,7 +10,12 @@ export * from "./issue_relation";
 export * from "./issue_sub_issues";
 export * from "./activity/base";
 
-export type TLoader = "init-loader" | "mutation" | "pagination" | undefined;
+export type TLoader =
+  | "init-loader"
+  | "mutation"
+  | "pagination"
+  | "loaded"
+  | undefined;
 
 export type TGroupedIssues = {
   [group_id: string]: string[];
