+++ packages/types/src/issues/issue.d.ts
@@ -25,6 +25,7 @@ export type TBaseIssue = {
   parent_id: string | null;
   cycle_id: string | null;
   module_ids: string[] | null;
+  type_id: string | null;
 
   created_at: string;
   updated_at: string;
@@ -48,6 +49,8 @@ export type TIssue = TBaseIssue & {
   issue_link?: TIssueLink[];
   // tempId is used for optimistic updates. It is not a part of the API response.
   tempId?: string;
+  // sourceIssueId is used to store the original issue id when creating a copy of an issue. Used in cloning property values. It is not a part of the API response.
+  sourceIssueId?: string;
 };
 
 export type TIssueMap = {
