+++ packages/ui/src/dropdowns/custom-search-select.tsx
@@ -15,6 +15,7 @@ export const CustomSearchSelect = (props: ICustomSearchSelectProps) => {
     customButtonClassName = "",
     buttonClassName = "",
     className = "",
+    chevronClassName = "",
     customButton,
     placement,
     disabled = false,
@@ -105,17 +106,22 @@ export const CustomSearchSelect = (props: ICustomSearchSelectProps) => {
                 <button
                   ref={setReferenceElement}
                   type="button"
-                  className={`flex w-full items-center justify-between gap-1 rounded border-[0.5px] border-custom-border-300 ${
-                    input ? "px-3 py-2 text-sm" : "px-2 py-1 text-xs"
-                  } ${
-                    disabled
-                      ? "cursor-not-allowed text-custom-text-200"
-                      : "cursor-pointer hover:bg-custom-background-80"
-                  } ${buttonClassName}`}
+                  className={cn(
+                    "flex w-full items-center justify-between gap-1 rounded border-[0.5px] border-custom-border-300",
+                    {
+                      "px-3 py-2 text-sm": input,
+                      "px-2 py-1 text-xs": !input,
+                      "cursor-not-allowed text-custom-text-200": disabled,
+                      "cursor-pointer hover:bg-custom-background-80": !disabled,
+                    },
+                    buttonClassName
+                  )}
                   onClick={toggleDropdown}
                 >
                   {label}
-                  {!noChevron && !disabled && <ChevronDown className="h-3 w-3 flex-shrink-0" aria-hidden="true" />}
+                  {!noChevron && !disabled && (
+                    <ChevronDown className={cn("h-3 w-3 flex-shrink-0", chevronClassName)} aria-hidden="true" />
+                  )}
                 </button>
               </Combobox.Button>
             )}
