+++ packages/ui/src/dropdowns/helper.tsx
@@ -12,6 +12,7 @@ export interface IDropdownProps {
   label?: string | JSX.Element;
   maxHeight?: "sm" | "rg" | "md" | "lg";
   noChevron?: boolean;
+  chevronClassName?: string;
   onOpen?: () => void;
   optionsClassName?: string;
   placement?: Placement;
