+++ web/ce/components/issue-types/values/update.tsx
@@ -1,0 +1,8 @@
+type TIssueExtraPropertyValuesUpdateProps = {
+  issueId: string;
+  issueTypeId: string;
+  projectId: string;
+  workspaceSlug: string;
+};
+
+export const IssueExtraPropertyValuesUpdate: React.FC<TIssueExtraPropertyValuesUpdateProps> = () => <></>;
