+++ web/ce/components/issues/index.ts
@@ -1,2 +1,3 @@
 export * from "./bulk-operations";
 export * from "./worklog";
+export * from "./issue-modal";
