+++ web/ce/components/issues/issue-modal/index.ts
@@ -1,0 +1,3 @@
+export * from "./form";
+export * from "./draft-issue-layout";
+export * from "./modal";
