+++ web/ce/constants/project/settings/index.ts
@@ -1,1 +1,2 @@
 export * from "./features";
+export * from "./tabs";
