+++ web/core/components/dropdowns/member/index.tsx
@@ -42,6 +42,7 @@ export const MemberDropdown: React.FC<Props> = observer((props) => {
     placement,
     projectId,
     showTooltip = false,
+    showUserDetails = false,
     tabIndex,
     value,
     icon,
@@ -73,6 +74,26 @@ export const MemberDropdown: React.FC<Props> = observer((props) => {
   const dropdownOnChange = (val: string & string[]) => {
     onChange(val);
     if (!multiple) handleClose();
+  };
+
+  const getDisplayName = (value: string | string[] | null, showUserDetails: boolean, placeholder: string = "") => {
+    if (Array.isArray(value)) {
+      if (value.length > 0) {
+        if (value.length === 1) {
+          return getUserDetails(value[0])?.display_name || placeholder;
+        } else {
+          return showUserDetails ? `${value.length} members` : "";
+        }
+      } else {
+        return placeholder;
+      }
+    } else {
+      if (showUserDetails && value) {
+        return getUserDetails(value)?.display_name || placeholder;
+      } else {
+        return placeholder;
+      }
+    }
   };
 
   return (
@@ -120,11 +141,7 @@ export const MemberDropdown: React.FC<Props> = observer((props) => {
               {!hideIcon && <ButtonAvatars showTooltip={showTooltip} userIds={value} icon={icon} />}
               {BUTTON_VARIANTS_WITH_TEXT.includes(buttonVariant) && (
                 <span className="flex-grow truncate text-xs leading-5">
-                  {Array.isArray(value) && value.length > 0
-                    ? value.length === 1
-                      ? getUserDetails(value[0])?.display_name
-                      : ""
-                    : placeholder}
+                  {getDisplayName(value, showUserDetails, placeholder)}
                 </span>
               )}
               {dropdownArrow && (
