+++ web/core/components/dropdowns/member/types.d.ts
@@ -7,6 +7,7 @@ export type MemberDropdownProps = TDropdownProps & {
   placeholder?: string;
   tooltipContent?: string;
   onClose?: () => void;
+  showUserDetails?: boolean;
 } & (
     | {
         multiple: false;
