+++ web/core/components/estimates/radio-select.tsx
@@ -4,7 +4,7 @@ import { cn } from "@/helpers/common.helper";
 
 type RadioInputProps = {
   name?: string;
-  label: string | React.ReactNode | undefined;
+  label?: string | React.ReactNode;
   wrapperClassName?: string;
   fieldClassName?: string;
   buttonClassName?: string;
@@ -46,14 +46,14 @@ export const RadioInput = ({
 
   return (
     <div className={className}>
-      <div className={cn(`mb-2`, inputLabelClassName)}>{inputLabel}</div>
+      {inputLabel && <div className={cn(`mb-2`, inputLabelClassName)}>{inputLabel}</div>}
       <div className={cn(`${wrapperClass}`, inputWrapperClassName)}>
         {options.map(({ value, label, disabled }, index) => (
           <div
             key={index}
             onClick={() => !disabled && setSelected(value)}
             className={cn(
-              "flex items-center gap-2",
+              "flex items-center gap-2 text-base",
               disabled ? `bg-custom-background-200 border-custom-border-200 cursor-not-allowed` : ``,
               inputFieldClassName
             )}
@@ -62,7 +62,7 @@ export const RadioInput = ({
               id={`${name}_${index}`}
               name={name}
               className={cn(
-                `group flex size-5 items-center justify-center rounded-full border border-custom-border-400 bg-custom-background-500 cursor-pointer`,
+                `group flex flex-shrink-0 size-5 items-center justify-center rounded-full border border-custom-border-400 bg-custom-background-500 cursor-pointer`,
                 selected === value ? `bg-custom-primary-200 border-custom-primary-100 ` : ``,
                 disabled ? `bg-custom-background-200 border-custom-border-200 cursor-not-allowed` : ``,
                 inputButtonClassName
@@ -72,7 +72,7 @@ export const RadioInput = ({
               disabled={disabled}
               checked={selected === value}
             />
-            <label htmlFor={`${name}_${index}`} className="text-base cursor-pointer">
+            <label htmlFor={`${name}_${index}`} className="cursor-pointer">
               {label}
             </label>
           </div>
