+++ web/core/components/issues/peek-overview/properties.tsx
@@ -28,6 +28,7 @@ import { getDate, renderFormattedPayloadDate } from "@/helpers/date-time.helper"
 import { shouldHighlightIssueDueDate } from "@/helpers/issue.helper";
 import { useIssueDetail, useMember, useProject, useProjectState } from "@/hooks/store";
 // plane web components
+import { IssueExtraPropertyValuesUpdate } from "@/plane-web/components/issue-types/values";
 import { IssueWorklogProperty } from "@/plane-web/components/issues";
 
 interface IPeekOverviewProperties {
@@ -288,6 +289,15 @@ export const PeekOverviewProperties: FC<IPeekOverviewProperties> = observer((pro
           issueId={issueId}
           disabled={disabled}
         />
+
+        {issue.type_id && (
+          <IssueExtraPropertyValuesUpdate
+            issueId={issueId}
+            issueTypeId={issue.type_id}
+            projectId={projectId}
+            workspaceSlug={workspaceSlug}
+          />
+        )}
       </div>
     </div>
   );
