+++ web/ee/components/issue-types/values/index.ts
@@ -1,0 +1,1 @@
+export * from "./update";
