+++ web/ee/components/issue-types/values/update.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/issue-types/values/update";
