+++ web/ee/components/issues/issue-modal/modal.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/issues/issue-modal/modal";
