+++ web/ee/constants/project/settings/tabs.ts
@@ -1,0 +1,1 @@
+export * from "ce/constants/project/settings/tabs";
